requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
apscheduler>=3.9.1
asyncpg==0.30.0
bcrypt==3.2.0
celery[auth,librabbitmq,sqlalchemy]==5.5.3
celery-sqlalchemy-scheduler==0.3.0
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
fastapi==0.115.14
fuse-python; sys_platform != "win32" and sys_platform != "darwin"
git+https://github.com/libfuse/python-fuse; sys_platform == "darwin"
greenlet==3.2.3
h11==0.16.0
httpx
idna==3.10
lark==1.2.2
markdown==3.8.2
msgraph-sdk==1.37.0
parse-torrent-title
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
python-dotenv>=1.1.1
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tinydb
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
